+++ management-ui/src/components/Home.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { withStyles } from 'material-ui';
+import { withStyles } from '@material-ui/core';
 
 const styles = theme => ({
   mainContent: {